Move utils under helpers
