+++ ci/kubetest/test_clickhouse_custom_settings_profile.py
@@ -2,7 +2,7 @@ import json
 
 import pytest
 
-from utils import cleanup_k8s, install_chart, kubectl_exec, wait_for_pods_to_be_ready
+from helpers.utils import cleanup_k8s, install_chart, kubectl_exec, wait_for_pods_to_be_ready
 
 VALUES_YAML = """
 cloud: local
