+++ ci/kubetest/test_clickhouse_persistence_disabled.py
@@ -1,7 +1,7 @@
 import pytest
 
 from helpers.clickhouse import get_clickhouse_statefulset_spec
-from utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
+from helpers.utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
 
 HELM_INSTALL_CMD = """
 helm upgrade \
