+++ ci/kubetest/test_clickhouse_persistence_enabled.py
@@ -2,7 +2,7 @@ import pytest
 from kubernetes import client
 
 from helpers.clickhouse import get_clickhouse_statefulset_spec
-from utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
+from helpers.utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
 
 HELM_INSTALL_CMD = """
 helm upgrade \
