+++ ci/kubetest/test_clickhouse_persistence_enabled_existing_claim.py
@@ -4,7 +4,7 @@ import pytest
 from kubernetes import client
 
 from helpers.clickhouse import get_clickhouse_statefulset_spec
-from utils import NAMESPACE, cleanup_k8s, exec_subprocess, helm_install, wait_for_pods_to_be_ready
+from helpers.utils import NAMESPACE, cleanup_k8s, exec_subprocess, helm_install, wait_for_pods_to_be_ready
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger()
