+++ ci/kubetest/test_grafana.py
@@ -1,6 +1,6 @@
 import pytest
 
-from utils import NAMESPACE, cleanup_k8s, install_chart, is_posthog_healthy, wait_for_pods_to_be_ready
+from helpers.utils import NAMESPACE, cleanup_k8s, install_chart, is_posthog_healthy, wait_for_pods_to_be_ready
 
 VALUES_YAML = """
 cloud: local
