+++ ci/kubetest/test_posthog_hpa_enabled.py
@@ -1,6 +1,6 @@
 import pytest
 
-from utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
+from helpers.utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
 
 HELM_INSTALL_CMD = """
 helm upgrade \
