+++ ci/kubetest/test_prometheus_postgres_exporter.py
@@ -1,7 +1,7 @@
 import pytest
 
 from helpers.metrics import is_prometheus_exporter_healthy
-from utils import cleanup_k8s, install_chart, is_posthog_healthy, wait_for_pods_to_be_ready
+from helpers.utils import cleanup_k8s, install_chart, is_posthog_healthy, wait_for_pods_to_be_ready
 
 VALUES_YAML = """
 cloud: local
