+++ ci/kubetest/test_redis_external.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from utils import (
+from helpers.utils import (
     cleanup_k8s,
     create_namespace_if_not_exists,
     helm_install,
