+++ ci/kubetest/test_redis_external_with_password.py
@@ -3,7 +3,7 @@ import logging
 
 import pytest
 
-from utils import (
+from helpers.utils import (
     cleanup_k8s,
     create_namespace_if_not_exists,
     helm_install,
